[MOV] sale: move/split account files (3)
